backend-django/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.2
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.0
gunicorn==21.2.0
whitenoise==6.5.0
django-extensions==3.2.3